Updates Capstone to latest version, removes access patches

+++ src/A64Architecture.cc
@@ -36,7 +36,7 @@ uint8_t A64Architecture::predecode(const void* ptr, uint8_t bytesAvailable,
     rawInsn.detail = &rawDetail;
 
     size_t size = 4;
-    size_t address = 0;
+    uint64_t address = 0;
 
     // TODO: capture result (success state) and replace instruction with an
     // "invalid decoding" implementation if not successful
